Change merge_asof direction to 'nearest' and remove tolerance argument
Tolerance argument does not work as expected when used by a colleague. Docs indicate that integer and timeseries values are the only appropriate things for tolerance. We're merging on floating point values that are the same between the two dfs (from the interpolation), so setting direction to nearest should account for floating point errors